% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chm.R
\name{chm}
\alias{chm}
\title{Simple Canopy Height Model}
\usage{
chm(las.path = NA, las.proj = NA, las.reproj = NA, nx = 100, ny = 100,
  fun = fun, silent = FALSE, plots = FALSE, geoTIFF = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{las.proj}{Proj4 projection string to use for projection. Defaults to NA.}

\item{las.reproj}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{nx}{Number of pixels along the x-axis. For a 50m radius plot, nx=100 is 1m resolution. Defaults to 100.}

\item{ny}{Number of pixels along the y-axis. For a 50m radius plot, ny=100 is 1m resolution. Defaults to 100.}

\item{fun}{Function for the calculate of height values in each cell. Defaults to max.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}

\item{plots}{Boolean switch for the saving of plot files to the las.path folder. Defaults to FALSE.}

\item{geoTIFF}{Boolean switch for the saving of projected GeoTIFF files to the las.path folder. Defaults to FALSE.}
}
\value{
The results of \code{chm}
}
\description{
This function implements various return-ratio-based fractional cover metrics
}
\examples{
chm(las.path='C:/plot.las', las.proj='+init=epsg:26911', las.reproj=NA, nx=100, ny=100, fun=max, silent=FALSE, plots=FALSE, geoTIFF=FALSE)
chm(las.path='C:/plot.las', las.proj='+init=epsg:26911', las.reproj=NA, nx=100, ny=100, fun=function(x) quantile(x, 0.95), silent=FALSE, plots=FALSE, geoTIFF=FALSE)
}
\keyword{canopy}
\keyword{chm,}
\keyword{height}
\keyword{model}

