% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chm.sf.R
\name{chm.sf}
\alias{chm.sf}
\title{Barycentric interpolation-based Spike-free Canopy Height Model}
\usage{
chm.sf(las = NA, las.proj = NA, las.reproj = NA, breaks = c(0.1, 0.25,
  0.5, 0.75), percent = TRUE, nx = 100, ny = 100, ko = 2.5, ku = 20,
  stacked = FALSE, silent = FALSE, plots = FALSE, geoTIFF = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{las.proj}{Proj4 projection string to use for projection. Defaults to NA.}

\item{las.reproj}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{breaks}{List of breaks to use for vertical stratification, either values or percentages of maximum height. Defaults to percentages: 0.10, 0.25, 0.50, 0.75.}

\item{percent}{Boolean switch for the use of percentages. Defaults to TRUE.}

\item{nx}{Number of pixels along the x-axis. For a 50m radius plot, nx=100 is 1m resolution. Defaults to 100.}

\item{ny}{Number of pixels along the y-axis. For a 50m radius plot, ny=100 is 1m resolution. Defaults to 100.}

\item{ko}{Barycentric triangle edge length filter for overstory layers. Defaults to 2.5.}

\item{ku}{Barycentric triangle edge length filter for the understory layer. Defaults to 20.}

\item{stacked}{Boolean switch for the output of a single raster or a raser stack. Defaults to FALSE.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}

\item{plots}{Boolean switch for the saving of plot files to the las.path folder. Defaults to FALSE.}

\item{geoTIFF}{Boolean switch for the saving of projected GeoTIFF files to the las.path folder. Defaults to FALSE.}
}
\value{
The results of \code{chm.sf}
}
\description{
This function implements a 2-D barycentric interpolation-based spike-free canopy height model algorithm
}
\examples{
chm.sf(las='C:/plot.las', las.proj='+init=epsg:26911', las.reproj=NA, breaks=c(0.10,0.25,0.50,0.75), percent=TRUE, nx=100, ny=100, ko=2.5, ku=20, stacked=FALSE, silent=FALSE, plots=FALSE, geoTIFF=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S0303243416300873}

\url{http://www.riegl.com/uploads/tx_pxpriegldownloads/khosravipour_SilviLaser2013.pdf}
}
\keyword{barycentric}
\keyword{canopy}
\keyword{chm,}
\keyword{height}
\keyword{interpolation}
\keyword{model,}
\keyword{spike-free}

