% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.crown.R
\name{dd.crown}
\alias{dd.crown}
\title{Nearest Crown Distance and Direction from Plot Center}
\usage{
dd.crown(crowns = NA, thresh.val = 1, degrees = FALSE, from = FALSE)
}
\arguments{
\item{crowns}{Raster object of tree crown positions. Defaults to NA.}

\item{thresh.val}{Threshold value used for tree crown selection. Defaults to 1.}

\item{degrees}{Boolean switch for the output of direction values in degrees rather than radians. Defaults to FALSE.}

\item{from}{Boolean switch for the output of direction values from nearest crowns rather than to nearest crowns. Defaults to FALSE.}
}
\value{
The results of \code{dd.crown}
}
\description{
This function calculates the distance and direction to the nearest crown pixel from the plot center
}
\examples{
dd.crown(crowns=itc.crowns, thresh.val=1, degrees=FALSE, from=FALSE)
}
\keyword{crown,}
\keyword{direction}
\keyword{distance,}

