% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.bl.R
\name{fc.bl}
\alias{fc.bl}
\title{Beer's Law-modified Intensity-return Ratio of Fractional Cover}
\usage{
fc.bl(las.path = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{fc.bl}
}
\description{
This function calculates fractional cover per the Beer's Law-modified intensity-return ratio, also known as the Intensity Cover Index
}
\examples{
fc.bl(las.path='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

