% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.p.R
\name{fc.p}
\alias{fc.p}
\title{Canopy-to-total-pixel Ratio}
\usage{
fc.p(chm = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{chm}{Name of the CHM raster object output from a CHM function with stacked=FALSE. Defaults to NA.}

\item{thresh.val}{Threshold value used for minimum canopy height. Defaults to 1.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{fc.p}
}
\description{
This function calculates the ratio of canopy pixels to non-canopy pixels
}
\examples{
fc.p(chm=chm, thresh.val=1.25, silent=FALSE)
}
\keyword{canopy,}
\keyword{pixel}
\keyword{ratio,}

