% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.r.R
\name{fc.r}
\alias{fc.r}
\title{Canopy-to-total-return Ratio Fractional Cover}
\usage{
fc.r(las.path = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{fc.r}
}
\description{
This function calculates fractional cover per the canopy-to-total-return ratio
}
\examples{
fc.r(las.path='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

