% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.hv.par.R
\name{gf.hv.par}
\alias{gf.hv.par}
\title{Parallel Hemispherical-Voronoi Gap Fraction}
\usage{
gf.hv.par(las.files = NA, models = "all", thresh.vals = seq(1, 4, 0.5),
  thresh.var = "height", reprojection = NA, pol.deg = 5, azi.deg = 45,
  col = "height", silent = TRUE, plots = FALSE)
}
\arguments{
\item{las.files}{String vector or list of LAS files. Defaults to NA.}

\item{models}{String vector or list of hemispherical lens models to use. Options include equi-distant (\code{"equidist"}), equi-angular (\code{"equiangle"}), stereographic (\code{"stereo"}), orthographic (\code{"ortho"}), or \code{"all"}. Defaults to \code{"all"}.}

\item{thresh.vals}{String vector or list of height thresholds to use. Defaults to 2.}

\item{thresh.var}{Specifies the LiDAR metric to use for thresholding canopy points. Options include height, intensity, nreturn, and class. Defaults to height.}

\item{reprojection}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{pol.deg}{Specifies the polar resolution for the radial plot lines. Defaults to 5.}

\item{azi.deg}{Specifies the azimuthal resolution for the radial plot lines. Defaults to 45.}

\item{col}{Specifies the LiDAR metric to use to color points of first plot in display. Options include height, intensity, nreturn, and class. Defaults to height.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{gf.hv.par}
}
\description{
This function allows you to run \code{gf.hv} on multiple plots in parallel using multi-core CPUs with \code{snow}
}
\examples{
gf.hv.par(las.files='C:/plot.las', models=c('equidist','stereo'), threshs=seq(1,4,0.5))
gf.hv.par(las.files=las.list, models='all', thresh.vals=1.25, thresh.var='height', reprojection=NA, pol.deg=5, azi.deg=45, col='height', silent=TRUE, plots=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
Forthcoming
}
\keyword{Voronoi,}
\keyword{fraction,}
\keyword{gap}
\keyword{hemispherical}
\keyword{parallel,}
\keyword{tesselation}

