% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.laie.aci.R
\name{gf.laie.aci}
\alias{gf.laie.aci}
\title{Volumentric-point-density-normalized Gap Fraction, Effective LAI, and ACI}
\usage{
gf.laie.aci(las.path = NA, pol.deg = 15, azi.deg = 45,
  reprojection = NA, silent = TRUE, plots = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{pol.deg}{Resolution of polar window in degrees. Defaults to 5.}

\item{azi.deg}{Resolution of azimuthal window in degrees. Defaults to 45.}

\item{reprojection}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to TRUE.}

\item{plots}{Boolean switch for the saving of plot files to the las.path folder. Defaults to FALSE.}
}
\value{
The results of \code{gf.laie.aci} in the form c(gf, e.lai, aci)
}
\description{
This function implements the volumentric-point-density-normalized gap fraction, effective LAI, and ACI algorithms
}
\examples{
gf.laie.aci(las.path='C:/plot.las', pol.deg=15, azi.deg=45, reprojection=NA, silent=FALSE, plots=FALSE)
}
\keyword{aci}
\keyword{fraction,}
\keyword{gap}
\keyword{lai,}

