% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itc.mw.R
\name{itc.mw}
\alias{itc.mw}
\title{Moving Window Individual Tree Crown Detection with Variable Window Size}
\usage{
itc.mw(chm = NA, res = 1, ht2rad = NA, min.h = 1, type = "circle",
  num = TRUE, silent = FALSE, plots = FALSE, geoTIFF = FALSE)
}
\arguments{
\item{chm}{Name of the CHM raster object output from a \code{chm} function with \code{stacked=FALSE}. Defaults to NA.}

\item{res}{Resolution of the input canopy height model in meters, used to adjust the ht2rad function. Defaults to 1.}

\item{ht2rad}{Function used to convert tree height to crown radius. Defaults to NA.}

\item{min.h}{Canopy minimum height. Defaults to 1.}

\item{type}{Shape parameter for the moving window function. Options are circle, Gauss, and rectangle. Defaults to circle.}

\item{num}{Boolean switch for the output of numeric values instead of a raster object. Defaults to TRUE.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}

\item{plots}{Boolean switch for the display of plots. Defaults to FALSE.}

\item{geoTIFF}{Boolean switch for saving GeoTIFF raster of tree positions to a folder, if the chm parameter is a folder. Defaults to FALSE.}
}
\value{
The results of \code{itc.mw}
}
\description{
This function detects individual tree crowns using a moving window function with variable window sizes
}
\examples{
itc.mw(chm=chm, res=1, ht2rad=function(x) 0.15746*x/res, min.h=1, type='circle', num=TRUE, silent=FALSE, plots=FALSE, geoTIFF=FALSE)
}
\references{
\url{http://www.mdpi.com/2072-4292/4/4/950/htm}
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\keyword{itc,}
\keyword{moving}
\keyword{window}
