% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itc.wat.R
\name{itc.wat}
\alias{itc.wat}
\title{Watershed Segmentation Individual Tree Crown Detection}
\usage{
itc.wat(chm = NA, res = 1, ht2rad = NA, min.h = 1, tolerance = 0.1,
  num = TRUE, silent = FALSE, ws.plot = FALSE, geoTIFF = FALSE)
}
\arguments{
\item{chm}{Name of the CHM raster object output from a \code{chm} function with \code{stacked=FALSE}. Defaults to NA.}

\item{res}{Resolution of the input canopy height model in meters, used to adjust the ht2rad function. Defaults to 1.}

\item{ht2rad}{Function used to convert tree height to crown radius. Defaults to NA.}

\item{min.h}{Canopy minimum height. Defaults to 1.}

\item{tolerance}{The minimum height between an object's highest point and the point where it contacts another object. If the height is smaller than the tolerance, the object will be combined with its highest neighbor. Defaults to 0.1.}

\item{num}{Boolean switch for the output of numeric values instead of a raster object. Defaults to TRUE.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}

\item{ws.plot}{Boolean switch for the display of the watershed segmentation plot. Defaults to FALSE.}

\item{geoTIFF}{Boolean switch for saving GeoTIFF raster of tree positions to a folder, if the chm parameter is a folder. Defaults to FALSE.}
}
\value{
The results of \code{itc.wat}
}
\description{
This function detects individual tree crowns using watershed segmentation
}
\examples{
itc.wat(chm=chm, res=1, ht2rad=function(x) 0.15746*x/res, min.h=1, tolerance=0.1, num=TRUE, silent=FALSE, ws.plot=FALSE, geoTIFF=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\keyword{itc,}
\keyword{segmentation}
\keyword{watershed}

