% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.sci.R
\name{lai.sci}
\alias{lai.sci}
\title{Solberg's Cover Index of Fractional Cover}
\usage{
lai.sci(las.path = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{lai.sci}
}
\description{
This function calculates fraction cover per Solberg's Cover Index
}
\examples{
lai.sci(las.path='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

