% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Le.bl.R
\name{Le.bl}
\alias{Le.bl}
\title{Simple Effective LAI Estimation with Beer's Law, the Ground-to-total-returns Ratio, and a Spherical Leaf Angle Distribution}
\usage{
Le.bl(las = NA, k = 0.5, silent = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{k}{Specifies the leaf angle distribution to use. Defaults to 0.5 for spherical.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{Le.bl}
}
\description{
This function calculates effective LAI using Beer's Law with the ground-to-total-returns ratio and a spherical leaf angle distribution
}
\examples{
Le.bl(las='C:/plot.las', k=0.5, silent=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S0168192309000409}
}
\keyword{LAI,}
\keyword{Le,}
\keyword{area}
\keyword{effective}
\keyword{index}
\keyword{leaf}

