% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Le.n.R
\name{Le.n}
\alias{Le.n}
\title{Contact Frequency and Fractional Cover-based LAI}
\usage{
Le.n(las = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{Le.n}
}
\description{
This function calculates LAI as a function of the effective LAI and fractional cover per contact frequency
}
\examples{
Le.n(las='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S0034425706001751}
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\keyword{LAI,}
\keyword{Le,}
\keyword{area}
\keyword{effective}
\keyword{index}
\keyword{leaf}
