% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.hv.R
\name{P.hv}
\alias{P.hv}
\title{Hemispherical-Voronoi Gap Fraction}
\usage{
P.hv(las = NA, model = "equidist", thresh.val = 1.25,
  thresh.var = "height", reprojection = NA, pol.deg = 5, azi.deg = 45,
  col = "height", plots = TRUE, plots.each = FALSE, plots.save = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{model}{Hemispherical lens geometry model to use. Options include equi-distant (\code{"equidist"}), equi-angular (\code{"equiangle"}), stereographic (\code{"stereo"}), and orthographic (\code{"ortho"}). Defaults to \code{"equidist"}.}

\item{thresh.val}{Specifies the value to use for thresholding. Defaults to 1.25.}

\item{thresh.var}{Specifies the LiDAR metric to use for thresholding canopy points. Options include height, intensity, nreturn, and class. Defaults to height.}

\item{reprojection}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{pol.deg}{Specifies the polar resolution for the radial plot lines. Defaults to 5.}

\item{azi.deg}{Specifies the azimuthal resolution for the radial plot lines. Defaults to 45.}

\item{col}{Specifies the LiDAR metric to use to color points of first plot in display. Options include height, intensity, nreturn, and class. Defaults to height.}

\item{plots}{Boolean switch for the interactive display of plots. Defaults to FALSE.}

\item{plots.each}{Boolean switch for displaying individual of plots. Defaults to FALSE.}

\item{plots.save}{Boolean switch for the saving of plot files to the las.path folder. Defaults to FALSE.}
}
\value{
The results of \code{P.hv}
}
\description{
This function implements Erickson's hemispherical-Voronoi gap fraction algorithm with four common lens geometries: equi-distant, equi-angular, stereographic, and orthographic
}
\examples{
P.hv(las='C:/plot.las', model='equidist', thresh.val=1.25, thresh.var='height', reprojection=NA, pol.deg=5, azi.deg=45, col='height', plots=TRUE, plots.each=FALSE, plots.save=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
Forthcoming
}
\keyword{fraction,}
\keyword{gap}
\keyword{thiessen}
\keyword{voronoi,}

