% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P.pdn.R
\name{P.pdn}
\alias{P.pdn}
\title{Point-density-normalized Gap Fraction, Effective LAI, and ACI}
\usage{
P.pdn(las = NA, pol.deg = 5, azi.deg = 45, reprojection = NA,
  silent = TRUE, plots = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{pol.deg}{Resolution of polar window in degrees. Defaults to 5.}

\item{azi.deg}{Resolution of azimuthal window in degrees. Defaults to 45.}

\item{reprojection}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to TRUE.}

\item{plots}{Boolean switch for the saving of plot files to the las.path folder. Defaults to FALSE.}
}
\value{
The results of \code{P.pdn} in the form c(Ppdn, Le, ACI)
}
\description{
This function implements Erickson's point-density-normalized gap fraction along with effective LAI and ACI algorithms
}
\examples{
P.pdn(las='C:/plot.las', pol.deg=15, azi.deg=45, reprojection=NA, silent=FALSE, plots=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
Forthcoming

\url{http://www.sciencedirect.com/science/article/pii/S0168192310000328}
}
\keyword{aci}
\keyword{fraction,}
\keyword{gap}
\keyword{lai,}

