% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial.grid.hemi.R
\name{radial.grid.hemi}
\alias{radial.grid.hemi}
\title{Modified Radial Grid Function for Hemispherical Lens Geometries}
\usage{
radial.grid.hemi(labels = NA, label.pos = NULL, radlab = FALSE,
  radial.lim = NULL, start = 0, clockwise = FALSE, label.prop = 1.1,
  grid.pos = grid.pos, grid.col = "gray", grid.bg = "transparent",
  show.radial.grid = TRUE, model = NA, r = "theta")
}
\arguments{
\item{labels}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to NA.}

\item{label.pos}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to NULL.}

\item{radlab}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to FALSE.}

\item{radial.lim}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to NULL.}

\item{start}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to 0.}

\item{clockwise}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to FALSE.}

\item{label.prop}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to 1.1.}

\item{grid.pos}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to grid.pos.}

\item{grid.col}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to gray.}

\item{grid.bg}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to transparent.}

\item{show.radial.grid}{Parameter adopted from the radial.grid function of the plotrix package. Defaults to TRUE.}

\item{model}{Hemispherical lens geometry model to use. Options include equi-distant (\code{"equidist"}), equi-angular (\code{"equiangle"}), stereographic (\code{"stereo"}), and orthographic (\code{"ortho"}). Defaults to NA.}

\item{r}{Hemispherical lens geometry string to use if no \code{model} is specified. String functions for \code{r} must be of the form \code{r = "tan(theta/2)"}. Defaults to \code{"theta"}.}
}
\value{
The results of \code{radial.grid.hemi}
}
\description{
This function creates radial grid plots for four common hemispherical lens geometries: equi-distant (\code{"equidist"}), equi-angular (\code{"equiangle"}), stereographic (\code{"stereo"}), and orthographic (\code{"ortho"})
}
\examples{
radial.grid.hemi()
}
\keyword{grid,}
\keyword{hemispherical}
\keyword{radial,}

