% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.path.R
\name{sun.path}
\alias{sun.path}
\title{Modified Version of Solar Position Plots by Thomas Steiner}
\usage{
sun.path(location.name = "Jena", phi = 50.9106, lambda = 11.5672,
  out = "", type = "polar", lang = "en", locale = "English",
  ieqot = FALSE, summert = 0)
}
\arguments{
\item{location.name}{Name of the plot location. Defaults to}

\item{phi}{Latitude in degrees.}

\item{lambda}{Longitude in degrees.}

\item{out}{Output file type. Options include PNG, PDF, JPG, and SVG.}

\item{type}{Type of plot. Options include polar and cart.}

\item{lang}{Language to use. Options include de for German, en for English, it for Italian, and nl for Dutch.}

\item{locale}{Set localization for month name generation. Options include German_Austria, English, Italian, and Dutch.}

\item{ieqot}{Boolean switch for whether to take the equation of time into account.}

\item{summert}{Boolean switch used to set to summertime normal time.}
}
\value{
The results of \code{sun.path}
}
\description{
This function creates Cartesian and polar sun path plots for specific locations, useful for estimating understory light. CC-by-SA 2.5: https://creativecommons.org/licenses/by-sa/2.5/
}
\examples{
sun.path(location.name='Jena', phi=50.9106, lambda=11.5672, out='', type='polar', lang='en', locale='English', ieqot=FALSE, summert=0)
}
\keyword{position,}
\keyword{solar}
\keyword{sonnenstand}
\keyword{sun}

