% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcc.cv.R
\name{vcc.cv}
\alias{vcc.cv}
\title{Cartesian-Voronoi Vertical Canopy Cover}
\usage{
vcc.cv(las = NA, reprojection = NA, col = "height", col2 = NA,
  thresh.var = "height", thresh.val = 1.25, silent = TRUE,
  plots = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{reprojection}{Proj4 projection string to use for reprojection. Defaults to NA.}

\item{col}{Specifies the LiDAR metric to use to color points of first plot in display. Options include height, intensity, nreturn, and class. Defaults to height.}

\item{col2}{Specifies the LiDAR metric to use to color points of second plot in display. Defaults to NA.}

\item{thresh.var}{Specifies the LiDAR metric to use for thresholding canopy points. Options include height, intensity, nreturn, and class. Defaults to height.}

\item{thresh.val}{Specifies the value to use for thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}

\item{plots}{Boolean switch for the saving of plot files to the las.path folder. Defaults to FALSE.}
}
\value{
The results of \code{vcc.cv}
}
\description{
This function implements the Cartesian-Voronoi vertical canopy cover algorithm of Alexander (2013)
}
\examples{
vcc.cv(las='C:/plot.las', reprojection=NA, col='height', col2='intensity', thresh.var='height', thresh.val=1.25, silent=FALSE, plots=FALSE)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S0034425713000771}
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}
\keyword{vertical}

