% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcc.fci.R
\name{vcc.fci}
\alias{vcc.fci}
\title{First-echo Cover Index of Vertical Canopy Cover}
\usage{
vcc.fci(las = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las}{Path or name of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{vcc.fci}
}
\description{
This function calculates fractional cover per the First-echo Cover Index
}
\examples{
vcc.fci(las='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\references{
\url{http://geography.swan.ac.uk/silvilaser/papers/oral_papers/Forestry\%20Applications\%20\%26\%20Inventory/Holmgren.pdf}

\url{http://link.springer.com/chapter/10.1007\%2F978-94-017-8663-8_20}
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}
\keyword{vertical}
