% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcc.p.R
\name{vcc.p}
\alias{vcc.p}
\title{Canopy-to-total-pixel Ratio Vertical Canopy Cover}
\usage{
vcc.p(chm = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{chm}{Name of the CHM raster object output from a CHM function with stacked=FALSE. Defaults to NA.}

\item{thresh.val}{Threshold value used for minimum canopy height. Defaults to 1.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{vcc.p}
}
\description{
This function calculates canopy cover as the ratio of canopy pixels to non-canopy pixels
}
\examples{
vcc.p(chm=chm, thresh.val=1.25, silent=FALSE)
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S0924271614001646}

\url{http://treephys.oxfordjournals.org/content/early/2009/06/24/treephys.tpp042.full}
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}
\keyword{vertical}
