% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.fci.R
\name{fc.fci}
\alias{fc.fci}
\title{First-echo Cover Index of Fractional Cover}
\usage{
fc.fci(las.path = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{fc.fci}
}
\description{
This function calculates fractional cover per the First-echo Cover Index
}
\examples{
fc.fci(las.path='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

