% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.fr.R
\name{fc.fr}
\alias{fc.fr}
\title{Canopy-to-first-return Ratio Fractional Cover}
\usage{
fc.fr(las.path = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{fc.fr}
}
\description{
This function calculates fractional cover as the canopy-to-first-return ratio
}
\examples{
fc.fr(las.path='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

