% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.hv.par.R
\name{gf.hv.par}
\alias{gf.hv.par}
\title{Parallel Hemispherical-Voronoi Gap Fraction}
\usage{
gf.hv.par(las.files = NA, models = NA, threshs = NA)
}
\arguments{
\item{las.files}{List of LAS files. Defaults to NA.}

\item{models}{List of hemispherical lens models to use. Defaults to equidist.}

\item{threshs}{List of height thresholds to use. Defaults to 2.}
}
\value{
The results of \code{gf.hv.par}
}
\description{
This function allows you to run gapfraction in parallel on multi-core CPUs
}
\examples{
gf.hv.par(las.files='C:/plot.las', models=c('equidist','stereo'), threshs=seq(1,3))
gf.hv.par(las.files=las.list, models='equidist', threshs=2)
}
\keyword{gapfraction}

