% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.hv.par.R
\name{gf.hv.par}
\alias{gf.hv.par}
\title{Parallel Hemispherical-Voronoi Gap Fraction}
\usage{
gf.hv.par(las.files = NA, models = NA, threshs = NA)
}
\arguments{
\item{las.files}{String vector or list of LAS files. Defaults to NA.}

\item{models}{String vector or list of hemispherical lens models to use. Options include equi-distant (\code{"equidist"}), equi-angular (\code{"equiangle"}), stereographic (\code{"stereo"}), and orthographic (\code{"ortho"}). Defaults to \code{"equidist"}.}

\item{threshs}{String vector or list of height thresholds to use. Defaults to 2.}
}
\value{
The results of \code{gf.hv.par}
}
\description{
This function allows you to run \code{gf.hv} on multiple plots in parallel using multi-core CPUs with \code{snow}
}
\examples{
gf.hv.par(las.files='C:/plot.las', models=c('equidist','stereo'), threshs=seq(1,3))
gf.hv.par(las.files=las.list, models='equidist', threshs=2)
}
\author{
Adam Erickson, \email{adam.erickson@ubc.ca}
}
\references{
Forthcoming
}
\keyword{Voronoi,}
\keyword{fraction,}
\keyword{gap}
\keyword{hemispherical}
\keyword{parallel,}
\keyword{tesselation}

