% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai.e.R
\name{lai.e}
\alias{lai.e}
\title{Simple Effective LAI Estimation with Beer's Law and the Ground-to-total-returns Ratio}
\usage{
lai.e(las.path = NA, k = 0.5, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{k}{Specifies the leaf angle distribution to use. Defaults to 0.5 for spherical.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{lai.e}
}
\description{
This function calculates effective LAI using Beer's Law with the ground-to-total-returns ratio and a spherical leaf angle distribution
}
\examples{
lai.e(las.path='C:/plot.las', k=0.5, silent=FALSE)
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

