% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai.n.R
\name{lai.n}
\alias{lai.n}
\title{Contact Frequency and Fractional Cover-based LAI}
\usage{
lai.n(las.path = NA, thresh.val = 1.25, silent = FALSE)
}
\arguments{
\item{las.path}{Path of LAS file. Defaults to NA.}

\item{thresh.val}{Specifies the value to use for canopy height thresholding. Defaults to 1.25.}

\item{silent}{Boolean switch for the interactive display of plots. Defaults to FALSE.}
}
\value{
The results of \code{lai.n}
}
\description{
This function calculates LAI as a function of the effective LAI and fractional cover per contact frequency
}
\examples{
lai.n(las.path='C:/plot.las', thresh.val=1.25, silent=FALSE)
}
\keyword{canopy}
\keyword{cover}
\keyword{cover,}
\keyword{fractional}

